Updated message sent and added twilio integration for comment function
Now, instead of sending the text of the comment, the link to the
comment is sent. Also, twilio integration has been added. This means
that a text message is now sent when a user comments. This was done for
more instant notification, as one has to check their personal messages
manually. Having both is repetitive, so the user may choose which one
he/she wants to implement. Both these features will soon be added for
the submission function.